move params to own file, make reference
